move config string parsing into config package
