+++ cmd/server/main.go
@@ -2,7 +2,6 @@ package main
 
 import (
 	"fmt"
-	"io/ioutil"
 	"math/rand"
 	"os"
 	"os/signal"
@@ -156,16 +155,12 @@ func main() {
 }
 
 func getConfig(c *cli.Context) (*config.Config, error) {
-	confString, err := getConfigString(c.String("config"), c.String("config-body"))
-	if err != nil {
-		return nil, err
-	}
-
 	strictMode := true
 	if c.Bool("disable-strict-config") {
 		strictMode = false
 	}
-	conf, err := config.NewConfig(confString, strictMode, c)
+
+	conf, err := config.NewConfig(c, strictMode)
 	if err != nil {
 		return nil, err
 	}
@@ -243,17 +238,4 @@ func startServer(c *cli.Context) error {
 	}()
 
 	return server.Start()
-}
-
-func getConfigString(configFile string, inConfigBody string) (string, error) {
-	if inConfigBody != "" || configFile == "" {
-		return inConfigBody, nil
-	}
-
-	outConfigBody, err := ioutil.ReadFile(configFile)
-	if err != nil {
-		return "", err
-	}
-
-	return string(outConfigBody), nil
 }
