+++ test/webhook_test.go
@@ -105,7 +105,7 @@ func TestWebhooks(t *testing.T) {
 }
 
 func setupServerWithWebhook() (server *service.LivekitServer, testServer *webhookTestServer, finishFunc func(), err error) {
-	conf, err := config.NewConfig("", true, nil)
+	conf, err := config.NewConfig(nil, true)
 	if err != nil {
 		panic(fmt.Sprintf("could not create config: %v", err))
 	}
